requirements.txt
python-telegram-bot[job-queue]==21.4
python-dotenv==1.0.1
tzdata
